+++ packages/editor/web/src/RichContentEditor/RichContentEditor.jsx
@@ -20,7 +20,16 @@ import {
   getBlockInfo,
   getFocusedBlockKey,
 } from 'wix-rich-content-editor-common';
-import { AccessibilityListener, normalizeInitialState, getLangDir } from 'wix-rich-content-common';
+import {
+  calculateDiff,
+  getPostContentSummary,
+} from 'wix-rich-content-editor-common/src/Utils/draftUtils';
+import {
+  AccessibilityListener,
+  normalizeInitialState,
+  getLangDir,
+  Version,
+} from 'wix-rich-content-common';
 import styles from '../../statics/styles/rich-content-editor.scss';
 import draftStyles from '../../statics/styles/draft.rtlignore.scss';
 
@@ -97,7 +106,10 @@ class RichContentEditor extends Component {
       locale,
       anchorTarget,
       relValue,
-      helpers,
+      helpers: {
+        ...helpers,
+        onPluginAdd: (...args) => helpers.onPluginAdd?.(...args, Version.currentVersion),
+      },
       config,
       isMobile,
       setEditorState: this.setEditorState,
@@ -190,8 +202,14 @@ class RichContentEditor extends Component {
   }
 
   updateEditorState = editorState => {
-    this.setEditorState(editorState);
-    this.props.onChange && this.props.onChange(editorState);
+    const onPluginDelete = this.props.helpers?.onPluginDelete;
+    if (onPluginDelete) {
+      calculateDiff(this.state.editorState, editorState, (...args) =>
+        onPluginDelete(...args, Version.currentVersion)
+      );
+      this.setEditorState(editorState);
+      this.props.onChange && this.props.onChange(editorState);
+    }
   };
 
   getCustomCommandHandlers = () => ({
@@ -411,6 +429,11 @@ class RichContentEditor extends Component {
     }
   }
 }
+
+RichContentEditor.publish = async (postId, editorState = {}, callBack = () => true) => {
+  const postSummary = getPostContentSummary(editorState);
+  callBack({ postId, ...postSummary });
+};
 
 RichContentEditor.propTypes = {
   editorKey: PropTypes.string,